+++ tests/unit/test_transforms.py
@@ -13,6 +13,7 @@ import datumaro.util.mask_tools as mask_tools
 from datumaro.components.annotation import (
     AnnotationType,
     Bbox,
+    Caption,
     Ellipse,
     Label,
     LabelCategories,
@@ -23,10 +24,13 @@ from datumaro.components.annotation import (
     Polygon,
     PolyLine,
     RleMask,
+    Tabular,
+    TabularCategories,
 )
 from datumaro.components.dataset import Dataset
 from datumaro.components.dataset_base import DatasetItem
-from datumaro.components.media import Image
+from datumaro.components.errors import MediaTypeError
+from datumaro.components.media import Image, Table, TableRow
 
 from ..requirements import Requirements, mark_bug, mark_requirement
 
@@ -1214,3 +1218,154 @@ class ReindexAnnotationsTest:
         assert n_anns == len(
             {(item.id, ann.id) for item in fxt_dataset for ann in item.annotations}
         )
+
+
+import argparse
+
+from pandas.api.types import CategoricalDtype
+
+
+class AstypeAnnotationsTest(TestCase):
+    def setUp(self):
+        self.table = Table.from_list(
+            [{"nswprice": 0.076108, "class": "DOWN"}, {"nswprice": 0.060376, "class": "UP"}]
+        )
+        self.dataset = Dataset.from_iterable(
+            [
+                DatasetItem(
+                    id="0",
+                    subset="train",
+                    media=TableRow(table=self.table, index=0),
+                    annotations=[Tabular(values={"class": "DOWN"})],
+                ),
+                DatasetItem(
+                    id="1",
+                    subset="train",
+                    media=TableRow(table=self.table, index=1),
+                    annotations=[Tabular(values={"class": "UP"})],
+                ),
+            ],
+            categories={
+                AnnotationType.tabular: TabularCategories.from_iterable(
+                    [("class", CategoricalDtype(), {"DOWN", "UP"})]
+                )
+            },
+            media_type=TableRow,
+        )
+        self.table_caption = Table.from_list([{"nswprice": 0.076108}, {"nswprice": 0.060376}])
+        self.dataset_caption = Dataset.from_iterable(
+            [
+                DatasetItem(
+                    id="0",
+                    subset="train",
+                    media=TableRow(table=self.table_caption, index=0),
+                    annotations=[Tabular(values={"nswprice": 0.076108})],
+                ),
+                DatasetItem(
+                    id="1",
+                    subset="train",
+                    media=TableRow(table=self.table_caption, index=1),
+                    annotations=[Tabular(values={"nswprice": 0.060376})],
+                ),
+            ],
+            categories={},
+            media_type=TableRow,
+        )
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    def test_split_arg_valid(self):
+        # Test valid input with a single colon
+        assert transforms.AstypeAnnotations._split_arg("date:label") == [("date", "label")]
+
+        # Test valid input with multiple colons
+        assert transforms.AstypeAnnotations._split_arg("date:label,title:text") == [
+            ("date", "label"),
+            ("title", "text"),
+        ]
+
+        # Test invalid input with no colon
+        with pytest.raises(argparse.ArgumentTypeError):
+            transforms.AstypeAnnotations._split_arg("datelabel")
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    def test_media_type(self):
+        dataset = Dataset.from_iterable(
+            [
+                DatasetItem(
+                    id="1",
+                    subset="train",
+                    annotations=[
+                        Label(0, id=0),
+                    ],
+                ),
+            ],
+            categories={},
+        )
+
+        with self.assertRaises(MediaTypeError):
+            transforms.AstypeAnnotations(dataset)
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    def test_transform_annotation_type_label(self):
+        table = self.table
+        expected = Dataset.from_iterable(
+            [
+                DatasetItem(
+                    id="0",
+                    subset="train",
+                    media=TableRow(table=table, index=0),
+                    annotations=[Label(label=0)],
+                ),
+                DatasetItem(
+                    id="1",
+                    subset="train",
+                    media=TableRow(table=table, index=1),
+                    annotations=[Label(label=1)],
+                ),
+            ],
+            categories={
+                AnnotationType.label: LabelCategories.from_iterable(
+                    [("DOWN", "class"), ("UP", "class")]
+                )
+            },
+            media_type=TableRow,
+        )
+
+        dataset = self.dataset
+        result = transforms.AstypeAnnotations(dataset)
+
+        categories = result._categories.get(AnnotationType.label, None)
+        assert categories
+
+        # Check label_groups of categories
+        assert categories.label_groups[0].name == "class"
+        assert sorted(categories.label_groups[0].labels) == ["DOWN", "UP"]
+
+        compare_datasets(self, expected, result)
+
+    @mark_requirement(Requirements.DATUM_GENERAL_REQ)
+    def test_transform_annotation_type_caption(self):
+        table = self.table_caption
+        expected = Dataset.from_iterable(
+            [
+                DatasetItem(
+                    id="0",
+                    subset="train",
+                    media=TableRow(table=table, index=0),
+                    annotations=[Caption("0.076108")],
+                ),
+                DatasetItem(
+                    id="1",
+                    subset="train",
+                    media=TableRow(table=table, index=1),
+                    annotations=[Caption("0.060376")],
+                ),
+            ],
+            categories={},
+            media_type=TableRow,
+        )
+
+        dataset = self.dataset_caption
+        result = transforms.AstypeAnnotations(dataset)
+
+        compare_datasets(self, expected, result)